aws_scripts/requirements.txt
python-dotenv==0.20.0
openpyxl==3.1.2
boto3==1.26.80
pylint==2.15.2
pyyaml==6.0.1